Create image_color_picker.py
When you run this code, a window will appear with an "Open Image" button. When you click the button, you can choose an image file. After loading the image, the window will display it, and you can hover the mouse over the image to see the corresponding hexadecimal color code displayed below the image. Note that the code works with common image formats like PNG, JPEG, and GIF. The image should have sufficient size to be displayed on the window properly.

Please keep in mind that Tkinter provides a simple GUI interface, and more advanced image processing or color picking features may require the use of additional libraries or frameworks.